Add basic implementation of AlexNet
This does not split the data across 2 GPUs as described in the paper, and does
not yet have Local Response Normalization layer, but has the overall structure.

We can't train or test this model yet due to no access to ImageNet, and there
does not appear to be an easily accessible alternative labeled datasets with the
same number of categories to match this model structure, so we'll leave this
as-s for now.