Add links to Colab and Binder; add use of einops
If these notebooks are viewed on GitHub, they are view-only and cannot be run.
Although we provide links in the README to open them in Colab and Binder, folks
who get a direct link to a notebook from outside of the repo, or folks who
quickly open a notebook may not have the links to open the notebook in those
services.

Make it trivial to open a runnable version of each notebook by embedding buttons
to open a notebook in either service.

Also, use `einops` for a more intuitive way to expand dimensions of the image
inputs rather than an unintuitive syntax of `np.expand_dims()`.