Combine notebook badges into a single cell
With large tables, as the number of cells in a row increases, the badges have
less space and are shrunk to fit, which reduces their readability. By combining
them all into a single cell, they will wrap, taking up the space of a single
badge horizontally, which will keep them readable with larger numbers of cells
per row.