Build GoogLeNet SVG diagrams in CI
Ensure that the diagrams continue to build buildable. Also install `m4` and
`dot` as they are necessary prerequisites for building these diagrams, and clear
out the directory of all SVGs before running the build.

The SVGs are currently checked-in so that we can display them properly with the
stylesheet on the published website; not deleting them would require that the
timestamps on the `*.m4` files would be newer than the SVG files, which we
cannot gaurantee, since the checked-in SVG files would have been created after
the relevant input files, so this step would never actually run in the CI
environment.