Remove unneeded fields from Jupyter notebooks
* Remove fields `base_uri`, `execution_count`, and `outputId`
* Set the field `id` to `null` for simplicity
* Delete all empty lists and dicts

This makes it easier to avoid more noise in the diffs as we update notebooks
over time with new features, refactorings, or simply cleanups.

Ensured that the notebooks are still openable by Colab; in particular, removing
the `id` field from elements makes them non-loadable, so that field is critical,
but its value can be set to `null` and Colab will assign a random value to it,
so we can clear it to avoid extra diffs.