Restore fields required by `nbconvert` and GitHub
Restored `metadata`, `outputs` and `execution_count` fields which are necessary
for the `nbconvert` tool and likely also required by the GitHub HTML preview.

Also restored the `id` field in the datasets exploration notebook which is
required for Colab to be able to load it.

Updated the cleanup script to properly handle the above.