Simplify GoogLeNet via Sequential pass-through
Create a new pseudo-layer `SequentialPassthrough` which is really just a for
loop that calls the underlying layers directly, rather than itself a layer like
`Sequential` actually is.

This allows full transparency into the underlying layers backing a model, while
simplifying how the code is wrtten and avoiding issues with passing in the wrong
inputs into the function calls.

The end result shows that the output of `model.summary()` has not changed at all
while the code has become cleaner and easier to read.