LeNet: add activation after MaxPooling2D layer
This layer does not accept an `activation` parameter as Conv2D and our custom
Subsampling layer do, so it needs to be added explicitly via an Activation
layer.

Also switch from using string shortcut names for activations to symbolic names for
consistency, and rename `lenet_activation` to `scaled_tanh` as it's more
descriptive.

Combine the activation functions definitions with the model architecture cell
for improved readability.