LeNet: reuse local Subsampling file if present
The `subsampling.py` already exists in the current directory, so if running this
notebook locally, just reuse it instead of fetching it from GitHub. Also, if
running in Colab, this ensures we only fetch it the first time, and continue to
reuse it for the duration of the notebook session, so it it updates in the
meantime, it does not change the local environment (and saves running time).