Add overridable paths to scripts in Makefile
We'd like to reuse the notebook formatting and testing scripts in another repo,
where these scripts will be in a `third_party` subdirectory and the scripts will
be referenced from there, where they won't be in `./` relative to the root of
the repo.

The ability to override the path to the scripts will make it possible to reuse
these scripts elsewhere without any local modifications.