LeNet: convert C5 from Dense to Conv2D
C5 is meant to be a convolutional layer (hence the "C" in the name), and per the
paper, even though the it functions similarly to a dense (fully-connected)
layer, it's a convolutional layer as it would auto-adjust to have more
connections (larger feature maps) if the input images were larger, whereas a
dense (fully-connected) layer which would remain the same size.

Training and testing remain nearly the same and hence, were not updated.