Ignore all Jupyter notebooks in `third_party`
When running format tests or automatically cleaning up or simplifying notebooks,
ignore everything in the `third_party` directory as there are many notebooks
there that we do not version or care to update.

Of note is that Mypy and PyType tests were already avoiding the `third_party`
directory; this updates the other tests to match that behavior as well.

Also add `third_party/python/.gitignore` so that it's ignored by git as well as
VS Code (which complains of too many untracked files otherwise), so that we can
install Python modules into that directory without affecting our tests or tools.