LeNet: simplify notebooks via common libraries
Having created the LeNet model construction module as well as a module for the
scaled `tanh` activation function, we can now remove the manual LeNet model
construction and also remove all the extra TensorFlow and Keras imports as the
libraries can do all of this for us, and we just have a few lines of code to
download the libraries, import them, and call them with the right parameters.

The recently added implementation notes docs walk through the evolution of the
code, so we can skip the tedious repetition of the same code, which needs to be
updated 3 times (at least!) whenever we make a change or a cleanup.

Now, this can be done in just one (centralized) place.