Use native Python `list` and `tuple` types
Starting with Python 3.9, we can use native `list` and `tuple` types for type
annotations, rather than having to import and use `List` and `Tuple`. For
Python versions prior to 3.9, we need to `from __future__ import annotations` to
make this work.

Since we've recently added testing with Python 3.9 and 3.9, we'll ensure this
works in practice with the various typechecking tooling we're using before
merging this change.