AlexNet: properly handle existing LRN module file
Only download if there isn't a local file by that name (which can happen in a
hosted environment such as Colab, Kaggle, or Binder), but not when running
locally, since the file is in the same directory as the notebook.

Also, when downloading another implementation from `third_party`, rename the
file so it doesn't override the local file by the same name.