Add params for LeNet and MNIST module downloads
This makes it easier to test an in-progress PR (since it will use a separate
branch), as well as to pin the dependent files to a release branch or tag (once
we have those) to keep the notebooks stable in the face of changing code.

Updated notebook titles to better match the filenames such that when the updated
notebooks are re-downloaded from Colab, they more closely match local filenames.

No functional changes.