Combine Python vars and Bash vars cells
Since we only use the Python-defined vars in a single Bash cell for downloading
files, for simplicity, we can combine the two into a single `%%bash` cell rather
than define them in Python and import them in a `%%script env ... bash` cell.

This is still easy to modify in a single place, which still makes testing easy.

No functional changes.