Separate LeNet implementations by directory
Create separate directories for the Keras and PyTorch implementations and move
the relevant files, notebooks, notes into those subdirectories, and adjust all
relevant paths accordingly.

This makes it easier to see what belongs to which implementation vs. when
everything is in a single directory, requiring framework-named suffixes for each
file.