Update notebooks to simplify module loading
Move to a consistent model of using full paths and module name as the basename
of the path to load multiple modules using a single for loop.

Also fixed in-notebook links to Colab for LeNet notebooks since they moved when
we moved the to `lenet/keras` subdirectory to make room for the PyTorch version.